fix/manifest: issue with generated
